requirements.txt
asgiref==3.8.1
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
Django==5.1
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.31.2
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
psutil==6.0.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.1.0
requests==2.32.3
setuptools==73.0.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
zope.event==5.0
zope.interface==7.0.1
